[Issue-3297] refactor: refactor asset types
