+++ packages/extension-base/src/core/substrate/foreign-asset-pallet.ts
@@ -3,14 +3,8 @@
 
 import { ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import { _getAppliedExistentialDeposit } from '@subwallet/extension-base/core/substrate/assets-pallet';
+import { PalletAssetsAssetAccount } from '@subwallet/extension-base/core/substrate/types';
 import BigN from 'bignumber.js';
-
-export type PalletAssetsAssetAccount = {
-  balance: number | string,
-  status: 'Frozen' | 'Liquid' | 'Blocked',
-  reason: Record<string, unknown>,
-  extra: unknown
-}
 
 export function _getForeignAssetPalletTransferable (accountInfo: PalletAssetsAssetAccount | undefined, existentialDeposit: string, extrinsicType?: ExtrinsicType): string {
   const strictMode = !extrinsicType || ![ExtrinsicType.TRANSFER_TOKEN, ExtrinsicType.TRANSFER_BALANCE].includes(extrinsicType);
