+++ packages/extension-base/src/core/substrate/types.ts
@@ -1,0 +1,48 @@
+// Copyright 2019-2022 @subwallet/extension-base
+// SPDX-License-Identifier: Apache-2.0
+
+// https://crates.parity.io/frame_system/struct.AccountInfo.html
+// https://wiki.polkadot.network/docs/learn-account-balances
+export type FrameSystemAccountInfoV2 = Omit<FrameSystemAccountInfoV1, 'data'> & {
+  data: {
+    free: number,
+    reserved: number,
+    frozen: number,
+    flags: number
+  }
+}
+
+export type FrameSystemAccountInfoV1 = {
+  nonce: number,
+  consumers: number,
+  providers: number,
+  sufficients: number,
+  data: {
+    free: number | string,
+    reserved: number,
+    miscFrozen: number,
+    feeFrozen: number
+  }
+}
+
+export type FrameSystemAccountInfo = FrameSystemAccountInfoV1 | FrameSystemAccountInfoV2;
+
+export type OrmlTokensAccountData = {
+  free: number,
+  reserved: number,
+  frozen: number
+}
+
+export type PalletAssetsAssetAccount = {
+  balance: number | string,
+  status: 'Frozen' | 'Liquid' | 'Blocked',
+  reason: Record<string, unknown>,
+  extra: unknown
+}
+
+export type PalletNominationPoolsPoolMember = {
+  poolId: number,
+  points: number,
+  lastRecordedRewardCounter: number,
+  unbondingEras: Record<string, number>
+}
